Lab 5 Line node Line reading,detection,estimation MotorNode changes Sensor calibration

+++ include/nodes/kinematics.hpp
@@ -11,8 +11,8 @@ namespace algorithms {
     };
 
     struct Encoders{
-        int l; //left
-        int r; //right
+        u_int16_t l; //left
+        u_int16_t r; //right
     };
     struct Coordinates{ //Cartesian coordinates
         float x;
